room controler merge
