+++ Psagot/Controllers/RoomController.cs
@@ -1,4 +1,5 @@
 ﻿using BL;
+using Entities.DTO;
 using Entities.Models;
 using Microsoft.AspNetCore.Mvc;
 
@@ -15,6 +16,22 @@ namespace Psagot.Controllers
             public RoomController(IRoomBL roomBL)
             {
                 _roomBL = roomBL;
+            }
+
+            [HttpPost("AddRoom")]
+            public async Task<IActionResult> AddRoom([FromBody] RoomDTO roomDTO)
+            {
+                var result = await _roomBL.AddRoom(roomDTO);
+                if (result.Room == null) return BadRequest(result.ErrorMessage);
+                return Ok(result.Room);
+            }
+
+            [HttpPut("UpdatrRoom")]
+            public async Task<IActionResult> UpdateRoom([FromBody] RoomDTO roomDTO)
+            {
+                var result = await _roomBL.UpdateRoom(roomDTO);
+                if (result.Room == null) return BadRequest(result.ErrorMessage);
+                return Ok(result.Room);
             }
 
             [HttpGet("GetAllRooms")]
